// frontend/src/api/generated/models/reservationDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */
import type { ReservationPersonDto } from "./reservationPersonDto";

export interface ReservationDto {
  id: number;
  /** @nullable */
  factoryId?: number | null;
  factoryDisplayName: string;
  startTime: string;
  endTime: string;
  durationHours: number;
  personnel: ReservationPersonDto[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}